ajustes em classe e controller de livros

+++ src/controllers/bookController.js
@@ -1,3 +1,4 @@
+import error404 from "../errors/error404.js";
 import { author } from "../models/Author.js";
 import book from "../models/Book.js";
 
@@ -15,7 +16,11 @@ class bookController {
     const id = req.params.id;
     try {
       const bookFound = await book.findById(id);
-      res.status(200).json(bookFound);
+      if (bookFound !== null) {
+        res.status(200).json(bookFound);
+      } else {
+        next(new error404("Falha na requisição: ID do livro não encontrado"));
+      }
     } catch (error) {
       next(error);
     }
@@ -34,10 +39,16 @@ class bookController {
   }
 
   static async updateBook(req, res, next) {
-    const id = req.params.id;
     try {
-      await book.findByIdAndUpdate(id, req.body);
-      res.status(200).json({ messege: "Atualizado com sucesso" });
+      const id = req.params.id;
+      const bookFound = await book.findById(id);
+
+      if (bookFound !== null) {
+        await book.findByIdAndUpdate(id, req.body);
+        res.status(200).json({ messege: "Atualizado com sucesso" });
+      } else {
+        next(new error404("Falha na requisição: ID do livro não encontrado"));
+      }
     } catch (error) {
       next(error);
     }
@@ -46,8 +57,14 @@ class bookController {
   static async deleteBook(req, res, next) {
     const id = req.params.id;
     try {
-      await book.findByIdAndDelete(id);
-      res.status(200).json({ messege: "Excluido com sucesso" });
+      const bookFound = await book.findById(id);
+
+      if (bookFound !== null) {
+        await book.findByIdAndDelete(id);
+        res.status(200).json({ messege: "Excluido com sucesso" });
+      } else {
+        next(new error404("Falha na requisição: ID do livro não encontrado"));
+      }
     } catch (error) {
       next(error);
     }
